cmd/cgo: perform explicit conversion in _GoStringLen
_GoStringLen performs an implicit conversion from intgo to size_t.
Explicitly cast to size_t.

This change avoids warnings when using cgo with CFLAGS:
-Wconversion.

GitHub-Last-Rev: b5df1ac0c3c90360fa1d22c069e0f126e9f894d8
GitHub-Pull-Request: golang/go#27092
Reviewed-on: https://go-review.googlesource.com/129820
Run-TryBot: Ian Lance Taylor <iant@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>

+++ src/cmd/cgo/out.go
@@ -1432,7 +1432,7 @@ void *CBytes(_GoBytes_);
 void *_CMalloc(size_t);
 
 __attribute__ ((unused))
-static size_t _GoStringLen(_GoString_ s) { return s.n; }
+static size_t _GoStringLen(_GoString_ s) { return (size_t)s.n; }
 
 __attribute__ ((unused))
 static const char *_GoStringPtr(_GoString_ s) { return s.p; }
